Former-commit-id: d143fc262e80bbda1bc71df8feeb197ee3ffa17a

+++ project/Starship/planet.cpp
@@ -8,12 +8,12 @@ Vect2d Planet::getCentre()
     return this->centre;
 }
 
-double Planet::getRadius()
+float Planet::getRadius()
 {
     return this->radius;
 }
 
-double Planet::getMass()
+float Planet::getMass()
 {
     return this->mass;
 }
@@ -23,12 +23,12 @@ void Planet::setCentre(Vect2d centre)
     this->centre = centre;
 }
 
-void Planet::setRadius(double radius)
+void Planet::setRadius(float radius)
 {
     this->radius = radius;
 }
 
-void Planet::setMass(double mass)
+void Planet::setMass(float mass)
 {
     this->mass = mass;
 }
